part 4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -1,23 +1,41 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({})
+        .populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response, next) => {
+    const body = request.body
+    const user = request.user
+    body.user = user._id
+    const blog = new Blog(body)
 
     try {
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     } catch (exception) {
         next(exception)
     }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        return response.status(400).json({
+            error: 'blog doesnt exist'
+        })
+    }
+    if (user.id.toString() !== blog.user.toString()) {
+        return response.status(401).json({ error: 'only creator can delete a blog' })
+    }
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
